+++ objects/src/accounts/account_id_anchor.rs
@@ -1,4 +1,4 @@
-use crate::{block::BlockNumber, AccountError, BlockHeader, Digest, EMPTY_WORD};
+use crate::{block::BlockNumber, errors::AccountIdError, BlockHeader, Digest, EMPTY_WORD};
 
 // ACCOUNT ID ANCHOR
 // ================================================================================================
@@ -49,23 +49,17 @@ impl AccountIdAnchor {
     /// Returns an error if any of the anchor constraints are not met. See the [type
     /// documentation](AccountIdAnchor) for details.
     pub fn new(
-        anchor_block_number: impl Into<BlockNumber>,
+        anchor_block_number: BlockNumber,
         anchor_block_hash: Digest,
-    ) -> Result<Self, AccountError> {
-        let anchor_block_number = anchor_block_number.into();
+    ) -> Result<Self, AccountIdError> {
         if anchor_block_number.as_u32() & 0x0000_ffff != 0 {
-            return Err(AccountError::AssumptionViolated(format!(
-          "TODO: Make proper error: anchor block must be an epoch block, i.e. its block number must be a multiple of 2^{}",
-          BlockNumber::EPOCH_LENGTH_EXPONENT)));
+            return Err(AccountIdError::AnchorBlockMustBeEpochBlock);
         }
 
         let anchor_epoch = anchor_block_number.block_epoch();
 
         if anchor_epoch == u16::MAX {
-            return Err(AccountError::AssumptionViolated(format!(
-                "TODO: Make proper error: anchor epoch cannot be {}",
-                u16::MAX
-            )));
+            return Err(AccountIdError::AnchorEpochMustNotBeU16Max);
         }
 
         Ok(Self {
@@ -113,7 +107,7 @@ impl AccountIdAnchor {
 // ================================================================================================
 
 impl TryFrom<&BlockHeader> for AccountIdAnchor {
-    type Error = AccountError;
+    type Error = AccountIdError;
 
     /// Extracts the [`BlockHeader::block_num`] and [`BlockHeader::hash`] from the provided
     /// `block_header` and tries to convert it to an [`AccountIdAnchor`].