+++ objects/src/accounts/account_id_prefix.rs
@@ -11,7 +11,7 @@ use vm_processor::DeserializationError;
 use super::account_id;
 use crate::{
     accounts::{account_id::validate_prefix, AccountIdVersion, AccountStorageMode, AccountType},
-    AccountError,
+    errors::AccountIdError,
 };
 
 // ACCOUNT ID PREFIX
@@ -68,7 +68,7 @@ impl AccountIdPrefix {
     ///
     /// Returns an error if any of the ID constraints of the prefix are not met. See the
     /// [`AccountId`](crate::accounts::AccountId) type documentation for details.
-    pub fn new(prefix: Felt) -> Result<Self, AccountError> {
+    pub fn new(prefix: Felt) -> Result<Self, AccountIdError> {
         validate_prefix(prefix)?;
 
         Ok(AccountIdPrefix { prefix })
@@ -137,7 +137,7 @@ impl From<AccountIdPrefix> for Felt {
 impl From<AccountIdPrefix> for [u8; 8] {
     fn from(id: AccountIdPrefix) -> Self {
         let mut result = [0_u8; 8];
-        result[..8].copy_from_slice(&id.prefix.as_int().to_le_bytes());
+        result[..8].copy_from_slice(&id.prefix.as_int().to_be_bytes());
         result
     }
 }
@@ -152,23 +152,26 @@ impl From<AccountIdPrefix> for u64 {
 // ================================================================================================
 
 impl TryFrom<[u8; 8]> for AccountIdPrefix {
-    type Error = AccountError;
+    type Error = AccountIdError;
 
-    /// Tries to convert a byte array in little-endian order to an [`AccountIdPrefix`].
+    /// Tries to convert a byte array in big-endian order to an [`AccountIdPrefix`].
     ///
     /// # Errors
     ///
     /// Returns an error if any of the ID constraints of the prefix are not met. See the
     /// [`AccountId`](crate::accounts::AccountId) type documentation for details.
-    fn try_from(value: [u8; 8]) -> Result<Self, Self::Error> {
-        let element =
-            Felt::try_from(&value[..8]).map_err(AccountError::AccountIdInvalidFieldElement)?;
-        Self::new(element)
+    fn try_from(mut value: [u8; 8]) -> Result<Self, Self::Error> {
+        // Felt::try_from expects little-endian order.
+        value.reverse();
+
+        Felt::try_from(value.as_slice())
+            .map_err(AccountIdError::AccountIdInvalidPrefixFieldElement)
+            .and_then(Self::new)
     }
 }
 
 impl TryFrom<u64> for AccountIdPrefix {
-    type Error = AccountError;
+    type Error = AccountIdError;
 
     /// Tries to convert a `u64` into an [`AccountIdPrefix`].
     ///
@@ -178,13 +181,13 @@ impl TryFrom<u64> for AccountIdPrefix {
     /// [`AccountId`](crate::accounts::AccountId) type documentation for details.
     fn try_from(value: u64) -> Result<Self, Self::Error> {
         let element = Felt::try_from(value.to_le_bytes().as_slice())
-            .map_err(AccountError::AccountIdInvalidFieldElement)?;
+            .map_err(AccountIdError::AccountIdInvalidPrefixFieldElement)?;
         Self::new(element)
     }
 }
 
 impl TryFrom<Felt> for AccountIdPrefix {
-    type Error = AccountError;
+    type Error = AccountIdError;
 
     /// Returns an [`AccountIdPrefix`] instantiated with the provided field .
     ///
@@ -236,7 +239,7 @@ impl Deserializable for AccountIdPrefix {
     fn read_from<R: ByteReader>(source: &mut R) -> Result<Self, DeserializationError> {
         <[u8; 8]>::read_from(source)?
             .try_into()
-            .map_err(|err: AccountError| DeserializationError::InvalidValue(err.to_string()))
+            .map_err(|err: AccountIdError| DeserializationError::InvalidValue(err.to_string()))
     }
 }